Access individual fuzz test results

Access individual fuzz test results

fuzz_value(fr, index = NULL, ...)

fuzz_call(fr, index = NULL, ...)

Arguments

fr

fuzz_results object

index

The test index (by position) to access. Same as the results_index in the data frame returned by as.data.frame.fuzz_results.

...

Additional arguments must be named regex patterns that will be used to match against test names. The names of the patterns must match the function argument name(s) whose test names you wish to match.

Functions

fuzz_value(): Access the object returned by the fuzz test

fuzz_call(): Access the call used for the fuzz test